Move validator classes to own directory
