+++ server_src/validators/dataobjectreferencevalidator.js
@@ -1,7 +1,7 @@
 'use strict';
 var Scenario = require('./../models/scenario').model;
 var Fragment = require('./../models/fragment').model;
-var parseToOLC = require('./json').parseToOLC;
+var parseToOLC = require('./../helpers/json').parseToOLC;
 
 /**
  * A validator that checks validity of data object references
