+++ server_src/validators/soundnessvalidator.js
@@ -1,7 +1,7 @@
 'use strict';
 var Scenario = require('./../models/scenario').model;
-var parseToBPMNObject = require('./json').parseToBPMNObject;
-var parseToOLC = require('./json').parseToOLC;
+var parseToBPMNObject = require('./../helpers/json').parseToBPMNObject;
+var parseToOLC = require('./../helpers/json').parseToOLC;
 
 /**
  * An validator that checks a fragment for structural soundness.
